[Frontend] Relationship Polar Area widget refetch (#14448)
